payments fix && migration workaround

+++ db/migrate/20130812092456_plan_development_remove_client.rb
@@ -1,5 +1,7 @@
 class PlanDevelopmentRemoveClient < ActiveRecord::Migration
   def change
+   if PlanDevelopment.new.attributes.has_key?("client")
     remove_column :plan_developments, :client
+   end
   end
 end
